+++ src/deluge/gui/ui/browser/sample_browser.h
@@ -52,8 +52,8 @@ public:
 	bool importFolderAsKit();
 	bool importFolderAsMultisamples();
 	ActionResult timerCallback();
-	bool claimCurrentFile(int32_t mayDoPitchDetection = 1, int32_t mayDoSingleCycle = 1,
-	                      int32_t mayDoWaveTable = 1); // 0 means no. 1 means auto. 2 means yes definitely
+	bool claimCurrentFile(int32_t mayDoPitchDetection = 1, int32_t mayDoSingleCycle = 1, int32_t mayDoWaveTable = 1,
+	                      bool loadWithoutExiting = false); // 0 means no. 1 means auto. 2 means yes definitely
 	bool renderMainPads(uint32_t whichRows, RGB image[][kDisplayWidth + kSideBarWidth],
 	                    uint8_t occupancyMask[][kDisplayWidth + kSideBarWidth], bool drawUndefinedArea = true);
 	void exitAndNeverDeleteDrum();
@@ -79,6 +79,8 @@ private:
 	void drawKeysOverWaveform();
 	void autoDetectSideChainSending(SoundDrum* drum, Source* source, char const* fileName);
 	void possiblySetUpBlinking();
+
+	bool autoLoadEnabled;
 
 	bool currentlyShowingSamplePreview;